medical-backend/scripts/requirements.txt
mysql-connector-python==8.2.0
faker==22.0.0
